mooroon5-news-service/requirements.txt
tavily-python
newspaper4k
google-generativeai
resend
python-dotenv
fastapi[standard]
uvicorn
email-validator
